コア機能においていたasync.jsをライブラリのSuspense.jsに移動
